Fix windows again

+++ prost-build/src/lib.rs
@@ -1007,7 +1007,7 @@ impl Config {
                 path.as_ref()
                     .as_os_str()
                     .encode_wide()
-                    .flat_map(|c| [((c & 0xff00) >> 8) as u8, c & 0xff])
+                    .flat_map(|c| [((c & 0xff00) >> 8) as u8, (c & 0xff) as u8])
                     .collect::<Vec<_>>()
                     .into()
             }
